pyproject.toml
[tool.poetry]
name = "yaml-pydantic-db"
version = "0.1.0"
description = ""
authors = ["Bartosz Cholewa <bbcholewa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
pydantic = {extras = ["email"], version = "^2.7.1"}
ruff = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
